+++ Dados/interfaces/IProduto.cs
@@ -31,7 +31,7 @@ namespace Dados
         /// <param name="preco">variavel para o preco do produto</param>
         /// <param name="garantia">variavel para a garantia do produto</param>
         /// <returns>retorna true se for alterado uma propriedade do produto e false se nao</returns>
-        bool AlterarProduto(int id, int[] t, string nome, string categoria, int preco, int garantia);
+        bool AlterarProduto(int id, int[] t, string nome, string categoria, double preco, int garantia);
 
         /// <summary>
         /// Funcao para retirar um produto da loja
